split up fetch module
